Added hash benchmarking file and debug info
Added requirement for benchmark library, so it actually runs now. Also
added minor debug output, to see which version of ruby is being used
when running the tests. Added hash.rb to test hash functions as well,
although the test in it could be more of an array test, may move it in
the future.